+++ backend/accounts/serializers/RegisterSerializer.py
@@ -1,0 +1,36 @@
+from django.contrib.auth.password_validation import validate_password
+from rest_framework import serializers
+from accounts.models import MealsHelperUser
+
+
+class RegisterSerializer(serializers.ModelSerializer):
+    email = serializers.EmailField(required=True)
+    password = serializers.CharField(write_only=True, required=True, validators=[validate_password])
+    password2 = serializers.CharField(write_only=True, required=True)
+
+    class Meta:
+        model = MealsHelperUser
+        fields = ('username', 'password', 'password2', 'email', 'first_name', 'last_name', 'phone_number', 'avatar')
+        extra_kwargs = {
+            'first_name': {'required': True},
+            'last_name': {'required': True},
+        }
+
+    def validate(self, attrs):
+        if attrs['password'] != attrs['password2']:
+            raise serializers.ValidationError({"password": "Password fields didn't match."})
+
+        return attrs
+
+    def create(self, validated_data):
+        user = MealsHelperUser.objects.create_user(
+            username=validated_data['username'],
+            email=validated_data['email'],
+            password=validated_data['password'],
+            first_name=validated_data['first_name'],
+            last_name=validated_data['last_name'],
+            phone_number=validated_data.get('phone_number'),
+            avatar=validated_data.get('avatar'),
+        )
+
+        return user
